factor out syntax highlighter
