+++ src/kilo.h
@@ -46,10 +46,10 @@ struct editorConfig {
 };
 
 enum DIRECTION {
-  UP = 0,
-  DOWN = 1,
-  LEFT = 2,
-  RIGHT = 3
+  LEFT = 0,
+  RIGHT = 1,
+  UP = 2,
+  DOWN = 3
 };
 
 enum KEY_ACTION {
