+++ src/syntax_highlighter.h
@@ -1,0 +1,9 @@
+#ifndef KILO_SYNTAX_HIGHLIGHTER_H
+#define KILO_SYNTAX_HIGHLIGHTER_H
+
+#include "kilo.h"
+
+void editorUpdateSyntax(erow *raw);
+int editorSyntaxToColor(int hl);
+
+#endif
